+++ examples/cards.html
@@ -1095,7 +1095,9 @@ Berlin
             }
 
 
-            .inline-latex svg {
+
+
+            .latex svg {
                 fill: var(--color-fg-default);
                 transform: scale(1.3);
                 margin-left: 20px;
