+++ examples/escape_test.html
@@ -877,7 +877,9 @@
             }
 
 
-            .inline-latex svg {
+
+
+            .latex svg {
                 fill: var(--color-fg-default);
                 transform: scale(1.3);
                 margin-left: 20px;
