+++ examples/markdown_demo.html
@@ -363,20 +363,20 @@ Sample text here...
     </div>
     <h4 id="syntax-highlighting">Syntax highlighting</h4>
     <p>Specify the language after the code block opening fences, 'auto' is also accepted.<br/></p>
-    <div id="code-div_ee28694c" class="multiline">
+    <div id="code-div_96e04923" class="multiline">
       <div class="code-block-scrollbar-container">
         <div class="code-block-container">
-          <pre><code name="multiline-code-block" class="code-block" id="code-block_ee28694c">```js
-var foo = function (bar) {
-  return bar++;
-};
-```</code></pre>
+          <pre><code name="multiline-code-block" class="code-block" id="code-block_96e04923">  ```js
+  var foo = function (bar) {
+    return bar++;
+  };
+  ```</code></pre>
         </div>
       </div>
-      <div class="copy-notification" id="copy-notification_ee28694c">
+      <div class="copy-notification" id="copy-notification_96e04923">
         Copied!
       </div>
-      <button class="btn-copy" id="copy-button_ee28694c" data-clipboard-target="#code-block_ee28694c">Copy</button>
+      <button class="btn-copy" id="copy-button_96e04923" data-clipboard-target="#code-block_96e04923">Copy</button>
     </div>
     <div id="code-div_c08782b0" class="multiline">
       <div class="code-block-scrollbar-container">
@@ -662,7 +662,6 @@ Second Term
     </div>
     <dl><b><dt>First Term</dt></b><dd>This is the definition of the first term.</dd><br/><b><dt>Second Term</dt></b><dd>This is one definition of the second term.</dd><dd>This is another definition of the second term.</dd><br/></dl>
     <h2 id="task-lists">Task lists</h2>
-    <p><code class="inline">&lt;div&gt;</code><br/></p>
     <div id="code-div_00ba3b40" class="multiline">
       <div class="code-block-scrollbar-container">
         <div class="code-block-container">
@@ -1560,7 +1559,9 @@ $$\frac{1}{2\pi i} \oint_\gamma \frac{f(z)}{z-z_0} \, dz = f(z_0) \cdot \sum_{n=
             }
 
 
-            .inline-latex svg {
+
+
+            .latex svg {
                 fill: var(--color-fg-default);
                 transform: scale(1.3);
                 margin-left: 20px;
