+++ examples/read_me.html
@@ -1120,7 +1120,9 @@
             }
 
 
-            .inline-latex svg {
+
+
+            .latex svg {
                 fill: var(--color-fg-default);
                 transform: scale(1.3);
                 margin-left: 20px;
