+++ src/carddown_parser/mdparser/escape_sequences.py
@@ -1,4 +1,8 @@
 from collections import OrderedDict
+from .htmltree import HtmlNode, TextNode
+from ..config import get_config
+
+config = get_config()
 
 ESCAPE_SEQUENCES = { 
     r"\\": {
@@ -73,3 +77,21 @@ ESCAPE_SEQUENCES = {
 }
 
 ESCAPE_SEQUENCES = OrderedDict(ESCAPE_SEQUENCES)
+
+
+def escape_text(text: str):
+    if config.document.prerender_latex is True:
+        ESCAPE_SEQUENCES["\$"] = {"intermediate" : "!!!ESCAPE!DOLLARSIGN!!!", "display_text" : "$"}
+    for str, esc in ESCAPE_SEQUENCES.items():
+        text = text.replace(str, esc["intermediate"])
+    return text
+
+
+def unescape_text_in_tree(root: HtmlNode):
+    for node in root:
+        if isinstance(node, TextNode):
+            for str, esc in ESCAPE_SEQUENCES.items():
+                if node.has_parent_with_tag("code"):
+                    node.text = node.text.replace(esc["intermediate"], str)
+                else:
+                    node.text = node.text.replace(esc["intermediate"], esc["display_text"])
