+++ src/carddown_parser/static/dark.css
@@ -479,7 +479,9 @@ li.L9 {
 }
 
 
-.inline-latex svg {
+
+
+.latex svg {
     fill: var(--color-fg-default);
     transform: scale(1.3);
     margin-left: 20px;
