+++ src/carddown_parser/static/light.css
@@ -191,7 +191,7 @@ The above copyright notice and this permission notice shall be included in all c
     word-wrap: normal;
 }
 
-.inline-latex svg {
+.latex svg {
     fill: var(--color-fg-default);
     transform: scale(1.3);
     margin-left: 20px;
